+++ packages/cli/CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.1.19 (2026-07-28)
+
+### 🧱 Updated Dependencies
+
+- Updated @jaw.id/core to 1.1.1
+
+## 0.1.18 (2026-07-27)
+
+### 🧱 Updated Dependencies
+
+- Updated @jaw.id/core to 1.1.0
+
 ## 0.1.17 (2026-07-16)
 
 ### 🩹 Fixes